feat(settings): combined /settings/email with Accounts + Signatures tabs (#229)
Ship slice 3 of the Settings redesign (PRD #226):

- /settings/email is now a tabbed shell wrapping AccountsTab + SignaturesTab.
  Accounts is the default tab; ?tab=signatures jumps straight to Signatures.
- AccountsTab is the existing email-accounts page body, extracted verbatim
  into a co-located *-tab.tsx and renamed to a named export. The four #142
  behavior tests moved alongside it (accounts-tab.test.tsx) and keep passing.
- SignaturesTab is the existing /settings/signatures page body, extracted
  the same way.
- /settings/signatures is now a one-line server redirect to
  /settings/email?tab=signatures so bookmarks and stale links still work.
- settings-nav.ts: the "Email Accounts" entry becomes "Email", and the
  separate "Email Signatures" entry is removed. The renamed entry keeps
  /settings/email so the mobile dropdown highlights it correctly.

Caught at build: SettingsTabs uses useSearchParams() and Next 16 requires
that hook to sit inside a Suspense boundary so the rest of the page can
prerender statically. Wrapped once inside the SettingsTabs deep module so
every shell — /settings/jobs included — gets the behavior for free; the
fallback renders the same tab strip pinned to defaultTab.
