+++ src/lib/settings-nav.ts
@@ -6,7 +6,6 @@ import {
   Receipt,
   Users,
   Mail,
-  FileSignature,
   Bell,
   FileText,
   Download,
@@ -36,8 +35,7 @@ export const settingsNavItems: SettingsNavItem[] = [
   { href: "/settings/vendors", label: "Vendors", icon: Store },
   { href: "/settings/expense-categories", label: "Expense Categories", icon: Receipt },
   { href: "/settings/users", label: "Users & Crew", icon: Users },
-  { href: "/settings/email", label: "Email Accounts", icon: Mail },
-  { href: "/settings/signatures", label: "Email Signatures", icon: FileSignature },
+  { href: "/settings/email", label: "Email", icon: Mail },
   { href: "/settings/contract-templates", label: "Contract Templates", icon: FileText },
   { href: "/settings/contracts", label: "Contracts", icon: Send },
   { href: "/settings/notifications", label: "Notifications", icon: Bell },
